+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React = require('react')
 import PropTypes = require('prop-types')
 import invariant = require('invariant')
-import deepClone = require('@jsbits/deep-clone')
 import { IconMap, IonIconDefs, IonIconProps, IonIconSizes, Nullable } from '..'
 
 type Dict<T = any> = { [k: string]: T }
@@ -34,9 +33,12 @@ const isObject = (obj: any) => !!obj && typeof obj === 'object'
  * Simple `Object.assign`-like function
  */
 const assign = function<T extends Dict, U extends Dict> (dest: T, src: U) {
+  dest = dest || {}
   if (src) {
     Object.keys(src).forEach((k) => {
-      dest[k] = src[k]
+      dest[k] = isObject(src[k])
+        ? assign(isObject(dest[k]) ? dest[k] : {}, src[k])
+        : src[k]
     })
   }
   return dest as Merge<T & U>
@@ -105,7 +107,7 @@ export const addIcons = function (iconMap: Nullable<IconMap>) {
  */
 export const setDefaults = function (defaults: Nullable<IonIconDefs>) {
   invariant(isObject(defaults), 'The defaults must be an object.')
-  const defs = assign(_Conf.defs, deepClone(defaults))
+  const defs = assign(_Conf.defs, defaults)
 
   // check and format class names, if any
   if (defs.className) {
@@ -150,10 +152,10 @@ export class IonIcon extends React.PureComponent<IonIconProps> {
   expandAttr(opts: Dict, name: keyof typeof expandFrom, value: any) {
     const [p1, p2] = expandFrom[name]
 
-    if (opts[p1] !== UNDEF) {
+    if (opts[p1] === UNDEF) {
       opts[p1] = value
     }
-    if (opts[p2] !== UNDEF) {
+    if (opts[p2] === UNDEF) {
       opts[p2] = value
     }
   }
@@ -169,6 +171,7 @@ export class IonIcon extends React.PureComponent<IonIconProps> {
     // Example of precedence [fill, stroke] = troke:
     // fill = user.fill -> defs.fill -> user.color -> defs.color
     // even if any of them is `null`, it will overrite.
+    debugger
 
     if (props.color !== UNDEF) {
       this.expandAttr(props, 'color', props.color)
@@ -209,7 +212,7 @@ export class IonIcon extends React.PureComponent<IonIconProps> {
   }
 
   render() {
-    const opts = deepClone(this.props) as Merge<IonIconProps>
+    const opts = assign({}, this.props) as Merge<IonIconProps>
     const name = opts.name
     delete opts.name